New refactored git-friendly repo structure, refactored files, no more duplicated code due to FeedforwardTrainer subclassing
