+++ examples/gan/prepare_gan_images.py
@@ -11,7 +11,7 @@ real_dir = "photo_jpg"  # change this to your photo directory
 image_ext = ".jpg"
 
 # change this to your desired output directory
-out_dir = "/Users/nicholasjunge/workspaces/maiot/ce_project/images_mini_test"
+out_dir = "/Users/nicholasjunge/workspaces/maiot/ce_project/images_mini"
 
 # change this number to adjust the dataset size.
 NUM_IMGS = 2
@@ -37,7 +37,8 @@ def prepare_gan_image_dir(num=100):
             with open(img, "rb") as img_file:
                 monet_img_raw = img_file.read()
             monet_img = tf.io.decode_image(monet_img_raw).numpy()
-            metadata = dict(zip(["height", "width", "num_channels"], monet_img.shape))
+            metadata = dict(zip(["height", "width", "num_channels"],
+                                monet_img.shape))
             file_name = os.path.basename(img)
             copyfile(img, os.path.join(out_dir, file_name))
