+++ examples/gan/preprocessing/preprocessor.py
